backend/requirements.txt
Django==5.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1
psycopg2-binary==2.9.9
python-decouple==3.8
python-dotenv==1.0.1
PyJWT==2.8.0
requests==2.31.0
urllib3==2.0.7
cryptography==41.0.7
Pillow==10.1.0
gunicorn==21.2.0
whitenoise==6.6.0
dj-database-url==2.1.0
google-auth>=2.15.0